+++ src/main/java/com/commerce/core/point/vo/PointDto.java
@@ -2,10 +2,8 @@ package com.commerce.core.point.vo;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class PointDto {
     private Long memberSeq;
-    private BigDecimal point;
+    private Long point;
 }
